Refactored splash screen more.
